Fix Android playback when startFrame > endFrame
Everytime we played an animation with startFrame > endFrame, we
were reversing the animation speed. As so, it played correctly
for the first time, but incorrectly on the second (as it was
reversed back to the regular animation speed).

Now, we only reverse the animation speed if needed. Additionally
we restore the original animation speed if needed for
startFrame < endFrame.